calculate 'mask' and sanitize 'score' within loop
storing ALL calculated masks in a tensor can consume a lot of RAM,
about 4GB for ~9hr audio.
so we save calculating the mask until the loop.
because the mask is now calculated in the loop, we have to wait until after
calculating the mask in order to np.nan_to_num(score)